# reports/verify-61c0f29/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Microservice Code Studio",
    "description": "Microservice Code Studio - Autonomous Java 21 / Spring Boot 3.x Generator",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/sessions": {
      "get": {
        "tags": [
          "Sessions"
        ],
        "summary": "List Sessions",
        "description": "Lists recent generation sessions ordered by creation date descending.",
        "operationId": "list_sessions_api_v1_sessions_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GenerationSessionListItem"
                  },
                  "title": "Response List Sessions Api V1 Sessions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Sessions"
        ],
        "summary": "Create Generation Session",
        "description": "Triggers an autonomous generation session for an ingested specification.\n\nReads the caller's LLM credentials from the same headers every other route\nuses. Without them this route silently selected DETERMINISTIC mode and emitted\noffline templates while `/quick-start` emitted model output (feature 011\nfollow-up), so a client could not tell which product it had received.",
        "operationId": "create_generation_session_api_v1_sessions_post",
        "parameters": [
          {
            "name": "X-LLM-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Llm-Api-Key"
            }
          },
          {
            "name": "X-LLM-Provider",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Llm-Provider"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationSessionSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sessions/{session_id}/export": {
      "get": {
        "tags": [
          "Export & Publish"
        ],
        "summary": "Export Session Project",
        "description": "Downloads the complete synthesized project as a standalone ZIP archive.",
        "operationId": "export_session_project_api_v1_sessions__session_id__export_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}/publish": {
      "post": {
        "tags": [
          "Export & Publish"
        ],
        "summary": "Publish Session Project",
        "description": "Publishes the project to a dedicated Git feature branch using ephemeral credentials.",
        "operationId": "publish_session_project_api_v1_sessions__session_id__publish_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateSessionRequest": {
        "properties": {
          "specId": {
            "type": "string",
            "title": "Specid",
            "description": "UUID of ingested specification"
          }
        },
        "type": "object",
        "required": [
          "specId"
        ],
        "title": "CreateSessionRequest"
      },
...
      "GenerationSessionListItem": {
        "properties": {
          "sessionId": {
            "type": "string",
            "title": "Sessionid"
          },
          "specId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specid"
          },
          "specName": {
            "type": "string",
            "title": "Specname"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "phase": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SessionPhase"
              },
              {
                "type": "null"
              }
            ]
          },
          "currentLifecyclePhase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currentlifecyclephase",
            "default": "INITIAL"
          },
          "lifecycleMode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lifecyclemode",
            "default": "GUIDED_STEP"
          },
          "completionPercentage": {
            "type": "number",
            "title": "Completionpercentage",
            "default": 0.0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "sessionId",
          "specName",
          "status",
          "createdAt"
        ],
        "title": "GenerationSessionListItem"
      },
      "GenerationSessionSummary": {
        "properties": {
          "sessionId": {
            "type": "string",
            "title": "Sessionid"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "queuePosition": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queueposition",
            "default": 0
          },
          "streamUrl": {
            "type": "string",
            "title": "Streamurl"
          }
        },
        "type": "object",
        "required": [
          "sessionId",
          "status",
          "streamUrl"
        ],
        "title": "GenerationSessionSummary"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PublishRequest": {
        "properties": {
          "repositoryUrl": {
            "type": "string",
            "title": "Repositoryurl",
            "description": "Target Git repository URL"
          },
          "branchName": {
            "type": "string",
            "title": "Branchname",
            "description": "Feature branch name e.g. feature/001-order-service"
          },
          "gitToken": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gittoken",
            "description": "Ephemeral Personal Access Token (PAT)"
          },
          "commitMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commitmessage",
            "description": "Git commit message",
            "default": "feat: initial autonomous generation and verified test suite"
          }
        },
        "type": "object",
        "required": [
          "repositoryUrl",
          "branchName"
        ],
        "title": "PublishRequest"
      },
      "PublishResponse": {
        "properties": {
          "branchUrl": {
            "type": "string",
            "title": "Branchurl"
          },
          "commitHash": {
            "type": "string",
            "title": "Commithash"
          },
          "pullRequestUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pullrequesturl"
          },
          "branchName": {
            "type": "string",
            "title": "Branchname"
          }
        },
        "type": "object",
        "required": [
          "branchUrl",
          "commitHash",
          "branchName"
        ],
        "title": "PublishResponse"
      },
...
      "SessionPhase": {
        "type": "string",
        "enum": [
          "INITIALIZATION",
          "SCAFFOLDING",
          "CODE_GENERATION",
          "TEST_SYNTHESIS",
          "SANDBOX_BUILD",
          "TEST_EXECUTION",
          "SELF_REPAIR",
          "SELF_REPAIR_LOOP",
          "VERIFIED",
          "FAILED"
        ],
        "title": "SessionPhase"
      },
      "SessionStatus": {
        "type": "string",
        "enum": [
          "QUEUED",
          "RUNNING",
          "PAUSED",
          "COMPLETED",
          "BLOCKED",
          "CANCELLED"
        ],
        "title": "SessionStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}